UCI Angular2 hw4 Using Services
Create a service for calculating total, count and average price for a list of prices.
- Values are calculated in one method and returned as an object.
Create a Summary component that shows total, count, and average price for a single category of products.
Using the new component and service to calculate values display the following:
- In the left column under the form and discount display summaries for categories that have a short name (less or equal 6 characters).
- In the right column under the table of products and discount display summaries for categories that have a long name (greater than 6
characters).
- Above the form display total for all products.
When a new product is entered or removed, the summaries should be updated accordingly